requirements.txt
python-telegram-bot==13.7
sqlalchemy==1.4.23
python-dotenv==0.19.0
psycopg2-binary==2.9.1
redis==4.0.2
aiohttp
python-dateutil==2.8.2
matplotlib==3.4.3
wordcloud
pandas
openai==0.27.0
google-generativeai==0.3.1 
